backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlmodel==0.0.22
pydantic-settings==2.5.2
